ProviderSection renamed to ImageQualityAndProviderSection
